[Lang] Fix warning messages (#6716)
This PR fixes some warning messages in the official code. Also, forbid
committing `.ply` files generated by running the `diff_sph` example in
this repo.

+++ python/taichi/examples/features/gui/gui_image_io.py
@@ -10,7 +10,7 @@ pixel = ti.field(ti.u8, shape=(512, 512, 3))
 @ti.kernel
 def paint():
     for I in ti.grouped(pixel):
-        pixel[I] = ti.random() * 255
+        pixel[I] = ti.u8(ti.random() * 255)
 
 
 paint()
